requirements.txt
neo4j
beautifulsoup4
requests
google-generativeai
python-dotenv
pydantic